requirements.txt
appnope==0.1.3
asttokens==2.4.1
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
fonttools==4.44.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
matplotlib==3.8.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.1
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sklearn==0.0
stack-data==0.6.3
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.9
zipp==3.17.0
